backend corrected with custom responses and errors

+++ server/src/controllers/authLogin.controller.js
@@ -13,11 +13,14 @@ export const loginUser = asyncHandler(async (req, res) => {
   }
 
   const isPasswordValid = await user.comparePassword(password);
-  if (!isPasswordValid && user.username !== username) {
+  if (!isPasswordValid && user.username !== username.toLowerCase()) {
     throw new ApiError(401, "Invalid credentials");
   }
 
-  const allUserSessions = await Session.find({ userId: user.id });
+  const allUserSessions = await Session.find({ userId: user._id }).sort({
+    createdAt: 1,
+  });
+
   const sessionLimit = user.role === "Admin" || user.role === "Manager" ? 2 : 1;
   if (allUserSessions.length >= sessionLimit) {
     await allUserSessions[0].deleteOne();
@@ -33,27 +36,25 @@ export const loginUser = asyncHandler(async (req, res) => {
     maxAge: 1 * 24 * 60 * 60 * 1000,
   });
 
-  return res.status(200).json(new ApiResponse(200, null, "Login successful!"));
+  return res
+    .status(200)
+    .json(new ApiResponse({ userId: user.id }, "Login successful!"));
 });
 
 export const logoutUser = asyncHandler(async (req, res) => {
   const { sid } = req.signedCookies;
+  if (!sid) {
+    throw new ApiError(401, "Not authenticated");
+  }
+
   await Session.findByIdAndDelete(sid);
-  res.clearCookie("sid");
-  return res
-    .status(200)
-    .json(new ApiResponse(200, null, "Logged out successfully"));
-});
+  res.clearCookie("sid", {
+    httpOnly: true,
+    signed: true,
+    sameSite: "none",
+    secure: true,
+    path: "/",
+  });
 
-export const logoutAllUserDevices = asyncHandler(async (req, res) => {
-  const { sid } = req.signedCookies;
-  const session = await Session.findById(sid);
-  await Session.deleteMany({ userId: session.userId });
-
-  res.clearCookie("sid");
-  return res
-    .status(200)
-    .json(
-      new ApiResponse(200, null, "Logged out from all devices successfully")
-    );
+  return res.status(200).json(new ApiResponse(null, "Logged out successfully"));
 });
